added backend of goals project node express mongodb

+++ backend/controllers/goalController.cjs
@@ -1,8 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const GoalModel = require('../model/goalModel.cjs')
+const UserModel = require('../model/userModel.cjs')
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await GoalModel.find()
+    
+    const goals = await GoalModel.find({ user: req.user.id })
     res.json(goals)
 })
 
@@ -14,7 +16,8 @@ const addGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add text')
     } else {
         const goal = await GoalModel.create({
-            text
+            text,
+            user: req.user.id
         })
         res.status(200).json(goal)
     }
@@ -31,6 +34,19 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await UserModel.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+
     const updatedGoal = await GoalModel.findByIdAndUpdate(id, req.body, { new: true })
 
     res.status(200).json(updatedGoal)
@@ -47,6 +63,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal Not Found')
     } else {
+
+        const user = await UserModel.findById(req.user.id)
+
+        if (!user) {
+            res.status(401)
+            throw new Error("User Not Found")
+        }
+
+        if (goal.user.toString() !== user.id) {
+            res.status(401)
+            throw new Error("User not authorized")
+        } 
+
+
         await GoalModel.findByIdAndDelete(id)
         res.status(200).json('Successfully Deleted')
     }
